Lock no Docker

+++ linux/12-docker/configure.sh
@@ -3,6 +3,9 @@
 echo " - Docker..."
 CONFS=$(dirname $0)/confs
 
+LOCK=/tmp/workstation.lock.docker
+[ -f ${LOCK} ] && return
+
 # Docker
 echo "deb [arch=amd64] https://download.docker.com/linux/ubuntu xenial stable" > /etc/apt/sources.list.d/docker.list
 apt-get update > /dev/null 2>&1
@@ -16,3 +19,5 @@ VERSION=1.21.0
 sudo curl -L https://github.com/docker/compose/releases/download/${VERSION}/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
 chmod 755 /usr/local/bin/docker-compose
 # Docker Compose
+
+> ${LOCK}
